requirements.txt
youtube_transcript_api
streamlit
google-generativeai>=0.7.0
python-dotenv
pathlib
argparse